move pgwire_number to codecs file
