+++ modules/module-postgres-storage/src/types/codecs.ts
@@ -88,3 +88,24 @@ export const hexBuffer = t.codec(
     return Buffer.from(encoded, 'hex');
   }
 );
+
+/**
+ * PGWire returns INTEGER columns as a `bigint`.
+ * This does a decode operation to `number`.
+ */
+export const pgwire_number = t.codec(
+  'pg_number',
+  (decoded: number) => decoded,
+  (encoded: bigint | number) => {
+    if (typeof encoded == 'number') {
+      return encoded;
+    }
+    if (typeof encoded !== 'bigint') {
+      throw new Error(`Expected either number or bigint for value`);
+    }
+    if (encoded > BigInt(Number.MAX_SAFE_INTEGER) || encoded < BigInt(Number.MIN_SAFE_INTEGER)) {
+      throw new RangeError('BigInt value is out of safe integer range for conversion to Number.');
+    }
+    return Number(encoded);
+  }
+);
